Refactor snapshot code via Connector trait (#99)
* Refactor snapshot code via Connector trait

* fix: lint
