+++ ndc-test/src/snapshot.rs
@@ -1,6 +1,55 @@
-use std::{fs::File, path::Path};
+use std::{collections::hash_map::DefaultHasher, fs::File, hash::{Hash, Hasher}, path::{Path, PathBuf}};
 
-use crate::error::{Error, Result};
+use async_trait::async_trait;
+use ndc_client::models;
+
+use crate::{connector::Connector, error::{Error, Result}};
+
+pub struct SnapshottingConnector<'a, C: Connector> {
+    pub snapshot_path: &'a PathBuf,
+    pub connector: &'a C,
+}
+
+#[async_trait(?Send)]
+impl <'a, C: Connector> Connector for SnapshottingConnector<'a, C> {
+    async fn get_capabilities(&self) -> Result<models::CapabilitiesResponse> {
+        let response: models::CapabilitiesResponse = self.connector.get_capabilities().await?;
+        snapshot_test(self.snapshot_path.join("capabilities").as_path(), &response)?;
+        Ok(response)
+    }
+
+    async fn get_schema(&self) -> Result<models::SchemaResponse> {
+        let response = self.connector.get_schema().await?;
+        snapshot_test(self.snapshot_path.join("schema").as_path(), &response)?;
+        Ok(response)
+    }
+
+    async fn query(&self, request: models::QueryRequest) -> Result<models::QueryResponse> {
+        let request_json = serde_json::to_string_pretty(&request).map_err(Error::SerdeError)?;
+        let response = self.connector.query(request).await?;
+
+        let mut hasher = DefaultHasher::new();
+        request_json.hash(&mut hasher);
+        let hash = hasher.finish();
+
+        let snapshot_subdir = {
+            let mut builder = self.snapshot_path.clone();
+            builder.extend(vec!["query", format!("{:x}", hash).as_str()]);
+            builder
+        };
+
+        snapshot_test(snapshot_subdir.join("expected.json").as_path(), &response)?;
+
+        std::fs::write(snapshot_subdir.join("request.json").as_path(), request_json)
+            .map_err(Error::CannotOpenSnapshotFile)?;
+
+        Ok(response)
+    }
+
+    async fn mutation(&self, request: models::MutationRequest) -> Result<models::MutationResponse> {
+        self.connector.mutation(request).await
+    }
+}
 
 pub fn snapshot_test<R>(snapshot_path: &Path, expected: &R) -> Result<()>
 where
@@ -11,12 +60,8 @@ where
         let snapshot: R = serde_json::from_reader(snapshot_file).map_err(Error::SerdeError)?;
 
         if snapshot != *expected {
-            let expected_json =
-                serde_json::to_string_pretty(&expected).map_err(Error::SerdeError)?;
-            return Err(Error::ResponseDidNotMatchSnapshot(
-                snapshot_path.into(),
-                expected_json,
-            ));
+            let actual = serde_json::to_string_pretty(&expected).map_err(Error::SerdeError)?;
+            return Err(Error::ResponseDidNotMatchSnapshot(snapshot_path.to_path_buf(), actual));
         }
     } else {
         let parent = snapshot_path.parent().unwrap();
